requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
drf-spectacular>=0.28
psycopg[binary]>=3.2
